# apps/backend/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/workspaces/{workspace_id}/pages/{page_id}": {
            "get": {
                "description": "Get page by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pages"
                ],
                "summary": "Get page by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma separated list of fields to include. Available fields: parentPage, childPages, owner, tasks",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Page"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to page or workspace",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update page by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pages"
                ],
                "summary": "Update page by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Page object that needs to be updated",
                        "name": "page",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pagesRouter.updatePageBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Page"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to page or workspace or no access to update page",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete page by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pages"
                ],
                "summary": "Delete page by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Page"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "403": {
                        "description": "No access to page or workspace or no access to delete page",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errorHandlers.Error": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "pagesRouter.updatePageBody": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "postgres.Page": {
            "type": "object",
            "properties": {
                "childPages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Page"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/postgres.User"
                },
                "parentPage": {
                    "$ref": "#/definitions/postgres.Page"
                },
                "role": {
                    "$ref": "#/definitions/postgres.UserRole"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Task"
                    }
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/postgres.PageType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.PageType": {
            "type": "string",
            "enum": [
                "board",
                "text",
                "group"
            ],
            "x-enum-varnames": [
                "PageTypeBoard",
                "PageTypeText",
                "PageTypeGroup"
            ]
        },
        "postgres.Task": {
            "type": "object",
            "properties": {
                "assignee": {
                    "$ref": "#/definitions/postgres.User"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "page": {
                    "$ref": "#/definitions/postgres.Page"
                },
                "reporter": {
                    "$ref": "#/definitions/postgres.User"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emailVerified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastLogin": {
                    "type": "string"
                },
                "lastPasswordReset": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "postgres.UserRole": {
            "type": "string",
            "enum": [
                "owner",
                "admin",
                "member",
                "commentator",
                "guest"
            ],
            "x-enum-varnames": [
                "UserRoleOwner",
                "UserRoleAdmin",
                "UserRoleMember",
                "UserRoleCommentator",
                "UserRoleGuest"
            ]
        },
...
    }
}